Warn when processing a measure without sample data rather than just fail

+++ spec/spec_helper.rb
@@ -39,6 +39,14 @@ end
 def validate_measures(measure_dirs, loader)
   
    measure_dirs.each do |dir|
+      # check for sample data
+      fixture_dir = File.join('fixtures', 'measures', File.basename(dir))
+      patient_files = Dir.glob(File.join(fixture_dir, 'patients', '*.json'))
+      if patient_files.length==0
+        puts "Skipping #{dir}, no sample data in #{fixture_dir}"
+        next
+      end
+
       puts "Parsing #{dir}"
 
       loader.drop_collection('bundles')
@@ -50,7 +58,6 @@ def validate_measures(measure_dirs, loader)
       measures = loader.save_measure(dir, 'measures')
       
       # load db with sample patient records
-      patient_files = Dir.glob(File.join('fixtures', 'measures', File.basename(dir), 'patients', '*.json'))
       patient_files.each do |patient_file|
         patient = JSON.parse(File.read(patient_file))
         loader.save('records', patient)
